pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "receipt-scanner"
version = "2.0.0"
description = "OCR-powered handwritten receipt scanner with hybrid EasyOCR + Azure Document Intelligence"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "Mahankali Harshith"},
]
keywords = ["ocr", "receipt", "scanner", "easyocr", "azure", "fastapi"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Framework :: FastAPI",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "easyocr==1.7.2",
    "opencv-python==4.13.0.92",
    "numpy==2.4.2",
    "azure-ai-documentintelligence==1.0.2",
    "azure-core==1.38.2",
    "openpyxl==3.1.5",
    "fastapi==0.132.0",
    "uvicorn[standard]==0.41.0",
    "python-multipart==0.0.22",
    "Pillow==12.1.1",
    "python-dotenv==1.2.1",
    "prometheus-fastapi-instrumentator==7.0.2",
    "opentelemetry-api==1.29.0",
    "opentelemetry-sdk==1.29.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.29.0",
    "opentelemetry-instrumentation-fastapi==0.50b0",
    "opentelemetry-instrumentation-requests==0.50b0",
    "opentelemetry-instrumentation-sqlite3==0.50b0",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "pytest-cov==6.1.1",
    "httpx==0.28.1",
    "pre-commit>=3.6",
    "ruff>=0.8",
    "mypy>=1.13",
    "bandit>=1.8",
    "pip-audit>=2.7",
]
postgres = [
    "psycopg2-binary>=2.9",
]

[project.urls]
Homepage = "https://github.com/MahankaliHarshith/OCR-project"
Repository = "https://github.com/MahankaliHarshith/OCR-project"
Issues = "https://github.com/MahankaliHarshith/OCR-project/issues"

# ─── Ruff (linter + formatter) ──────────────────────────────────────────────
[tool.ruff]
target-version = "py312"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort (import sorting)
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "TCH",    # type-checking imports
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults (FastAPI Depends)
    "SIM108", # use ternary operator (readability preference)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

# ─── Pytest ──────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

# ─── Coverage ────────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["app"]
omit = [
    "app/__pycache__/*",
    "app/static/*",
    "app/ocr/engine.py",
    "app/ocr/azure_engine.py",
    "app/ocr/hybrid_engine.py",
    "app/ocr/preprocessor.py",
    "app/ocr/total_verifier.py",
    "app/ocr/quality_scorer.py",
    "app/ocr/usage_tracker.py",
    "app/ocr/image_cache.py",
    "app/ocr/validators.py",
    "app/services/receipt_service.py",
    "app/services/batch_service.py",
    "app/services/product_service.py",
    "app/training/*",
    "app/db_postgres.py",
    "app/error_tracking.py",
    "app/observability.py",
    "app/main.py",
    "app/api/routes.py",
]

[tool.coverage.report]
show_missing = true
skip_empty = true
fail_under = 70
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "if TYPE_CHECKING:",
    "pass",
]

# ─── Mypy (type checking) ───────────────────────────────────────────────────
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
check_untyped_defs = false
disallow_untyped_defs = false

# ─── Bandit (security linter) ───────────────────────────────────────────────
[tool.bandit]
exclude_dirs = ["tests", "scripts"]
skips = ["B101", "B601"]
